requirements.txt
openai
pytesseract
Pillow
